Don't accrue delegation rewards to the undelegated
This is remarkably hard to do for something so simple.

* We have to consider the last provider to know if we're going to or from delegation. (Rell has no window functions. Outer joins break update.)
* We also need to suppress rewards until start date.
* This is in three places.

It would be possible to substitute a look-back/ahead with a corralated sub-query if we had sequences, but these are not directly available in Rell.

As the code is in three places there is a test that pokes the internal model to make sure they don't diverge.